OGM-122 cassandra backend
split backend tests that use classes with mappings that cassandra can't support.
