+++ core/src/test/java/org/hibernate/ogm/backendtck/associations/manytoone/ManyToOneTest.java
@@ -13,8 +13,6 @@ import static org.hibernate.ogm.utils.TestHelper.getNumberOfEntities;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import java.util.Arrays;
-
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.ogm.utils.GridDialectType;
@@ -258,45 +256,6 @@ public class ManyToOneTest extends OgmTestCase {
 	}
 
 	@Test
-	public void testUnidirectionalOneToMany() throws Exception {
-		final Session session = openSession();
-		Transaction tx = session.beginTransaction();
-		Product beer = new Product( "Beer", "Tactical nuclear penguin" );
-		session.persist( beer );
-
-		Product pretzel = new Product( "Pretzel", "Glutino Pretzel Sticks" );
-		session.persist( pretzel );
-
-		Basket basket = new Basket();
-		basket.setId( "davide_basket" );
-		basket.setOwner( "Davide" );
-		basket.setProducts( Arrays.asList( beer, pretzel ) );
-		session.persist( basket );
-
-		tx.commit();
-		session.clear();
-
-		tx = session.beginTransaction();
-		basket = (Basket) session.get( Basket.class, basket.getId() );
-		assertThat( basket ).isNotNull();
-		assertThat( basket.getId() ).isEqualTo( basket.getId() );
-		assertThat( basket.getProducts() )
-			.onProperty( "name" ).containsOnly( beer.getName(), pretzel.getName() );
-		tx.commit();
-
-		session.clear();
-
-		tx = session.beginTransaction();
-		session.delete( basket );
-		session.delete( pretzel );
-		session.delete( beer );
-		tx.commit();
-		session.close();
-
-		checkCleanCache();
-	}
-
-	@Test
 	public void testDefaultBiDirManyToOneCompositeKeyTest() throws Exception {
 		Session session = openSession();
 		Transaction transaction = session.beginTransaction();
@@ -348,8 +307,6 @@ public class ManyToOneTest extends OgmTestCase {
 				SalesGuy.class,
 				Beer.class,
 				Brewery.class,
-				Basket.class,
-				Product.class,
 				Game.class,
 				Court.class
 		};
